Add BertScoreMax transform for qa_accuracy
