+++ src/fmeval/eval_algorithms/qa_accuracy.py
@@ -4,9 +4,7 @@ from functools import partial
 from typing import Any, Callable, Dict, List, Optional, Set, Union
 from dataclasses import dataclass
 
-import ray
 from nltk.metrics.scores import f_measure, precision, recall
-from ray.actor import ActorHandle
 
 from fmeval.constants import (
     BERTSCORE_DEFAULT_MODEL,
@@ -25,6 +23,7 @@ from fmeval.eval_algorithms import (
     EvalOutput,
     EvalScore,
 )
+from fmeval.transforms.common import BertScoreMax, BERT_SCORE
 from fmeval.model_runners.model_runner import ModelRunner
 from fmeval.transforms.transform import Transform
 from fmeval.transforms.transform_pipeline import TransformPipeline
@@ -37,14 +36,11 @@ from fmeval.util import (
     assert_condition,
 )
 
-# from fmeval.transforms.qa_accuracy_metrics import BertScoreWithDelimiter, BERT_SCORE
-
 F1_SCORE = "f1_score"
 EXACT_MATCH_SCORE = "exact_match_score"
 QUASI_EXACT_MATCH_SCORE = "quasi_exact_match_score"
 PRECISION_OVER_WORDS = "precision_over_words"
 RECALL_OVER_WORDS = "recall_over_words"
-BERT_SCORE = "bert_score"
 
 # for metrics that are included in the QAAccuracyScores Transform
 QA_ACCURACY_SCORE_NAMES = [
@@ -249,104 +245,6 @@ class QAAccuracyScores(Transform):
         return record
 
 
-class BertScoreWithDelimiter(Transform):
-    """The abstract base class for QA Accuracy metric transforms.
-
-    Concrete subclasses of QAAccuracyMetric should simply implement the
-    `compute_metric` method and their own __init__ method. Subclasses need not implement
-    the __call__ method, as it is already implemented in this class, but are
-    free to do so if additional customization is required.
-    """
-
-    def __init__(
-        self,
-        target_output_keys: List[str],
-        model_output_keys: List[str],
-        output_keys: List[str],
-        allow_duplicate_input_keys: bool,
-        bertscore_model: Union[BertscoreHelperModel, ActorHandle],
-        target_output_delimiter: Optional[str] = "<OR>",
-    ):
-        """QAAccuracyMetric initializer.
-
-        Note that the ordering of the elements in `target_output_keys`, `model_output_keys`,
-        and `output_keys` must match, i.e. the kth element of `target_output_keys` and the
-        kth element of `model_output_keys` are used to compute the kth metric, which has
-        an output key of `output_keys[k]`.
-
-        :param target_output_keys: The keys corresponding to target outputs.
-        :param model_output_keys: The keys corresponding to model outputs.
-        :param output_keys: The output keys for this Transform, which correspond
-            to the metrics/scores that get computed.
-        :param allow_duplicate_input_keys: Whether to allow duplicate keys in
-            `target_output_keys` and `model_output_keys`. This parameter is usually
-            False, but will be True when a SummarizationAccuracyMetric is created
-            to compute metrics on perturbed model outputs. In this case,
-            `target_output_keys` will be a list of a single repeated key, while
-            `model_output_keys` will contain the keys for perturbed model outputs.
-        :param *args: Variable length argument list.
-        :param **kwargs: Arbitrary keyword arguments.
-        """
-        assert_condition(
-            len(target_output_keys) == len(model_output_keys) and len(target_output_keys) == len(output_keys),
-            "len(target_output_keys), len(model_output_keys) and len(output_keys) should all match. "
-            f"len(target_output_keys) is {len(target_output_keys)}, len(model_output_keys) is "
-            f"{len(model_output_keys)}, and len(output_keys) is {len(output_keys)}.",
-        )
-        super().__init__(
-            target_output_keys,
-            model_output_keys,
-            output_keys,
-            allow_duplicate_input_keys,
-            bertscore_model,
-            target_output_delimiter,
-        )
-        self.register_input_output_keys(
-            target_output_keys + model_output_keys,
-            output_keys,
-            allow_duplicates=allow_duplicate_input_keys,
-        )
-        self.target_output_keys = target_output_keys
-        self.model_output_keys = model_output_keys
-        self.bertscore_model = bertscore_model
-        self.target_output_delimiter = target_output_delimiter
-
-    @validate_call
-    def __call__(self, record: Dict[str, Any]) -> Dict[str, Any]:
-        """Augment the input record with metrics computed via self.compute_metric.
-
-        :param record: The input record.
-        :returns: The input record with metrics added in.
-        """
-        for target_output_key, model_output_key, output_key in zip(
-            self.target_output_keys, self.model_output_keys, self.output_keys
-        ):
-            score = self.compute_metric(record[target_output_key], record[model_output_key])
-            record[output_key] = score
-        return record
-
-    def compute_metric(self, target_output: str, model_output: str) -> float:
-        """Compute the metric that is specific to this Transform.
-
-        :param target_output: The target/reference output.
-        :param model_output: The actual output produced by the model.
-        :returns: A float representing the computed metric value.
-        """
-        possible_targets = target_output.split(self.target_output_delimiter)
-        if isinstance(self.bertscore_model, BertscoreHelperModel):
-            return max([self.bertscore_model.get_helper_scores(target, model_output) for target in possible_targets])
-        else:
-            possible_scores = list(
-                map(
-                    lambda x: self.bertscore_model.get_helper_scores.remote(x, model_output),  # type: ignore[return-value]
-                    possible_targets,
-                )
-            )
-            all_scores = ray.get(possible_scores)
-
-            return max(all_scores)
-
-
 @dataclass(frozen=True)
 class QAAccuracyConfig(EvalAlgorithmConfig):
     """Configures the QA Accuracy evaluation algorithm.
@@ -407,8 +305,11 @@ class QAAccuracy(EvalAlgorithmInterface):
         super().__init__(eval_algorithm_config)
 
         self.bertscore_model = BertscoreHelperModel(eval_algorithm_config.model_type_for_bertscore)
-        qa_accuracy_score = QAAccuracyScores(target_output_delimiter=eval_algorithm_config.target_output_delimiter)
-        bert_score = BertScoreWithDelimiter(
+
+        # Saving QAAccuracyScores in the original self.transform
+        self.transform = QAAccuracyScores(target_output_delimiter=eval_algorithm_config.target_output_delimiter)
+
+        bert_score = BertScoreMax(
             target_output_keys=[DatasetColumns.TARGET_OUTPUT.value.name],
             model_output_keys=[DatasetColumns.MODEL_OUTPUT.value.name],
             output_keys=[BERT_SCORE],
@@ -417,9 +318,8 @@ class QAAccuracy(EvalAlgorithmInterface):
             target_output_delimiter=eval_algorithm_config.target_output_delimiter,
         )
         self._eval_algorithm_config = eval_algorithm_config
-        self.qa_accuracy_score = qa_accuracy_score  # saving the QAAccuracyScores transform
         self.bert_score = bert_score  # saving the BertScore transform
-        self.pipeline = TransformPipeline([qa_accuracy_score, bert_score])
+        self.pipeline = TransformPipeline([self.transform, bert_score])
 
     def evaluate_sample(self, target_output: str, model_output: str) -> List[EvalScore]:
         """Compute QA accuracy metrics for a single sample.
@@ -465,17 +365,17 @@ class QAAccuracy(EvalAlgorithmInterface):
         """
         # Create a shared resource to be used during the evaluation.
         bertscore_shared_resource = create_shared_resource(self.bertscore_model)
-        bert_score = BertScoreWithDelimiter(
+        bert_score = BertScoreMax(
             target_output_keys=[DatasetColumns.TARGET_OUTPUT.value.name],
             model_output_keys=[DatasetColumns.MODEL_OUTPUT.value.name],
             output_keys=[BERT_SCORE],
             allow_duplicate_input_keys=True,
             bertscore_model=bertscore_shared_resource,
-            target_output_delimiter=self.qa_accuracy_score.target_output_delimiter,
+            target_output_delimiter=self.transform.target_output_delimiter,
         )
 
         # Create a new pipeline that uses the shared resource instead of self.bertscore_model.
-        pipeline = TransformPipeline([self.qa_accuracy_score, bert_score])
+        pipeline = TransformPipeline([self.transform, bert_score])
 
         dataset_configs = get_dataset_configs(dataset_config, self.eval_name)
         eval_outputs = []
